add StatusNotification request/response for OCPP 2.0
