+++ ocpp-json/src/main/scala/com/thenewmotion/ocpp/json/v20/Ocpp20Procedure.scala
@@ -108,6 +108,7 @@ object CsmsOcpp20Procedures extends Ocpp20Procedures[
   val procedures = List(
     CsmsOcpp20Procedure[BootNotificationRequest, BootNotificationResponse, CsmsReqRes],
     CsmsOcpp20Procedure[HeartbeatRequest, HeartbeatResponse, CsmsReqRes],
-    CsmsOcpp20Procedure[TransactionEventRequest, TransactionEventResponse, CsmsReqRes]
+    CsmsOcpp20Procedure[TransactionEventRequest, TransactionEventResponse, CsmsReqRes],
+    CsmsOcpp20Procedure[StatusNotificationRequest, StatusNotificationResponse, CsmsReqRes]
   )
 }
