+++ ocpp-messages/src/main/scala/com/thenewmotion/ocpp/messages/v20/ReqRes.scala
@@ -11,7 +11,8 @@ object CsReqRes {
 }
 
 object CsmsReqRes {
-  implicit object BootNotificationReqRes extends CsmsReqRes[BootNotificationRequest, BootNotificationResponse]
-  implicit object HeartbeatReqRes extends CsmsReqRes[HeartbeatRequest, HeartbeatResponse]
-  implicit object TransactionEventReqRes extends CsmsReqRes[TransactionEventRequest, TransactionEventResponse]
+  implicit object BootNotificationReqRes   extends CsmsReqRes[BootNotificationRequest  , BootNotificationResponse  ]
+  implicit object HeartbeatReqRes          extends CsmsReqRes[HeartbeatRequest         , HeartbeatResponse         ]
+  implicit object TransactionEventReqRes   extends CsmsReqRes[TransactionEventRequest  , TransactionEventResponse  ]
+  implicit object StatusNotificationReqRes extends CsmsReqRes[StatusNotificationRequest, StatusNotificationResponse]
 }
